Split bootkick and watchdog driver sources
